+++ Pl/t/deref-args-01.t
@@ -13,7 +13,6 @@ use warnings;
 
 use lib ".";
 
-use Data::Dump qw/dump/;
 
 use PPI;
 use PPI::Dumper;