+++ Pl/t/named-unary-01.t
@@ -14,7 +14,6 @@ use warnings;
 
 use lib ".";
 
-use Data::Dump qw/dump/;
 use PPI;
 use PPI::Dumper;
 use Test::More tests => 32;